Dont send empty artifact lists to delete during cleanup

+++ nephoria/testcases/ec2/ebs/legacy_ebs_test_suite.py
@@ -1110,9 +1110,12 @@ class LegacyEbsTestSuite(CliTestRunner):
         Attempts to clean up test artifacts created during this test
         """
         for zone in self.zonelist:
-            self.user.ec2.terminate_instances(zone.instances)
-            self.user.ec2.delete_volumes(zone.volumes)
-            self.user.ec2.delete_snapshots(self.snaps)
+            if zone.instances:
+                self.user.ec2.terminate_instances(zone.instances)
+            if zone.volumes:
+                self.user.ec2.delete_volumes(zone.volumes)
+            if self.snaps:
+                self.user.ec2.delete_snapshots(self.snaps)
         self.user.ec2.delete_keypair(self.keypair)
